Move init to zero test to separate file
