+++ autometrics/tests/integration_test.rs
@@ -190,20 +190,3 @@ fn build_info() {
         && line.contains(r#"service_name="autometrics""#)
         && line.ends_with("} 1")));
 }
-
-#[cfg(debug_assertions)]
-#[test]
-fn zero_metrics() {
-    prometheus_exporter::try_init().ok();
-
-    #[autometrics]
-    fn zero_metrics_fn() {}
-
-    // Note that we are not calling the function, but it should still be registered
-
-    let metrics = prometheus_exporter::encode_to_string().unwrap();
-    println!("{}", metrics);
-    assert!(metrics
-        .lines()
-        .any(|line| line.contains(r#"function="zero_metrics_fn""#) && line.ends_with("} 0")));
-}
